tests/wpt/tests/html/semantics/embedded-content/the-img-element/decode/image-decode-with-quick-attach.html

<!DOCTYPE html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>HTMLImageElement.prototype.decode(), attach to DOM before promise resolves.</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel=help href="https://html.spec.whatwg.org/multipage/embedded-content.html#dom-img-decode">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body></body>

<script>
"use strict";

let png = "/images/green.png?image-decode-with-quick-attach-" + Math.random();

function run_test(prop = "src") {
  const img = new Image();
  img[prop] = png;
  const promise = img.decode().then(function(arg) {
    assert_equals(arg, undefined);
  });
  // Don't wait for the promise to resolve before attaching the image.
  // The promise should still resolve successfully.
  document.body.appendChild(img);
  return promise;
}

promise_test(() => run_test(), document.title + ": src not cached");
promise_test(() => run_test(), document.title + ": src cached");
promise_test(() => run_test("srcset"), document.title + ": srcset");
</script>
